[CPDLP-1850] Move the new service classes to another module
