+++ spec/factories/services/npq/npq_application.rb
@@ -58,7 +58,7 @@ FactoryBot.define do
 
     trait :accepted do
       after :create do |npq_application|
-        NPQ::AcceptApplication.new(npq_application:).call
+        NPQ::Application::Accept.new(npq_application:).call
         npq_application.reload
       end
     end
