+++ spec/requests/api/v2/npq_applications_spec.rb
@@ -22,7 +22,7 @@ RSpec.describe "NPQ Applications API", :with_default_schedules, type: :request d
       list << create_list(:npq_application, 2, npq_lead_provider: other_npq_lead_provider, school_urn: "123456", npq_course:, cohort:)
 
       list.flatten.each do |npq_application|
-        NPQ::AcceptApplication.new(npq_application:).call
+        NPQ::Application::Accept.new(npq_application:).call
       end
     end
 
@@ -277,6 +277,7 @@ RSpec.describe "NPQ Applications API", :with_default_schedules, type: :request d
         post "/api/v2/npq-applications/#{npq_profile.id}/reject"
 
         expect(parsed_response.key?("errors")).to be_truthy
+        expect(parsed_response["errors"][0]["title"]).to eql("npq_application")
         expect(parsed_response.dig("errors", 0, "detail")).to eql("Once accepted an application cannot change state")
       end
     end
